show sshLogin Statistics enhance permissioncheck for groups/orgs

+++ app/Http/Controllers/GroupsController.php
@@ -137,6 +137,7 @@ class GroupsController extends Controller
     public function show(Group $group)
     {
         abort_unless(\Gate::allows('group_show'), 403);
+        abort_unless(auth()->user()->groups->contains($group), 403);
         LogController::set(get_class($this).'@'.__FUNCTION__, $group->id);
          // axios call?
         if (request()->wantsJson()){
@@ -152,6 +153,7 @@ class GroupsController extends Controller
     public function edit(Group $group)
     {
         abort_unless(\Gate::allows('group_edit'), 403);
+        abort_unless(auth()->user()->groups->contains($group), 403);
 
         $grades  = Organization::where('id', auth()->user()->current_organization_id)->get()->first()->type->grades()->get(); //Grade::all();
         $periods = Period::all();//Organization::where('id',auth()->user()->current_organization_id)->get()->first()->periods;
@@ -168,6 +170,7 @@ class GroupsController extends Controller
     public function update(UpdateGroupRequest $request, Group $group)
     {
         abort_unless(\Gate::allows('group_edit'), 403);
+        abort_unless(auth()->user()->groups->contains($group), 403);
 
         $group->update([
             'title' => $request['title'],
@@ -186,6 +189,7 @@ class GroupsController extends Controller
     public function destroy(Group $group)
     {
         abort_unless(\Gate::allows('group_delete'), 403);
+        abort_unless(auth()->user()->groups->contains($group), 403);
 
         // first delete all relations
         $group->curricula()->detach();
@@ -201,6 +205,8 @@ class GroupsController extends Controller
     public function massDestroy(MassDestroyGroupRequest $request)
     {
         abort_unless(\Gate::allows('group_delete'), 403);
+        abort_unless(auth()->user()->groups->contains($group), 403);
+
         Group::whereIn('id', request('ids'))->delete();
 
         return response(null, 204);
@@ -212,6 +218,8 @@ class GroupsController extends Controller
 
         foreach ((request()->enrollment_list) AS $enrolment)
         {
+            abort_unless(auth()->user()->groups->contains($enrolment['group_id']), 403);
+
             $group = Group::findOrFail($enrolment['group_id']);
             $user = User::findOrFail($enrolment['user_id']);
             //if user isn't enrolled to organization, enrol with student role
@@ -233,6 +241,7 @@ class GroupsController extends Controller
 
         foreach ((request()->expel_list) AS $expel)
         {
+            abort_unless(auth()->user()->groups->contains($expel['group_id']), 403);
             $user = User::find($expel['user_id']);
             $return[] = $user->groups()->detach($expel['group_id']);
         }
